* added x86_64 libs and dlls for cppunit * removed warnings uncovered by switching to 64 bit target, mainly involving a different-sized size_t * decreased warning level on pcre project

+++ SoarSuite/Core/CLI/src/cli_source.cpp
@@ -319,10 +319,10 @@ bool CommandLineInterface::StreamSource( std::istream& soarStream, const std::st
 		}
 
 		// Fire off the command
-		unsigned oldResultSize = m_Result.str().size();
+		size_t oldResultSize = m_Result.str().size();
 		if (DoCommandInternal(command)) {
 			// Add trailing newline if result changed size
-			unsigned newResultSize = m_Result.str().size();
+			size_t newResultSize = m_Result.str().size();
 			if (newResultSize > 0 && (oldResultSize != newResultSize)) {
 				// but don't add after sp's
 				if (m_Result.str()[m_Result.str().size()-1] != '*') {
